% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepNN.R
\docType{package}
\name{deepNN-package}
\alias{deepNN-package}
\alias{deepNN}
\title{deepNN}
\format{
An object of class \code{logical} of length 1.
}
\usage{
deepNN
}
\description{
Teaching resources (yet to be added) and implementation of some Deep Learning methods. Includes multilayer perceptron, different activation functions, regularisation strategies, stochastic gradient descent and dropout.
}
\details{
\tabular{ll}{
Package: \tab deepNN\cr
Version: \tab 0.1\cr
Date: \tab 2019-01-11\cr
License: \tab GPL-3 \cr
}



section{Dependencies}{
The package \code{deepNN} depends upon some other important contributions to CRAN in order to operate; their uses here are indicated:\cr\cr
    stats, graphics.
}

section{Citation}{
deepNN: Deep Learning. Benjamin M. Taylor
}

references{
Thanks go to the following references for helping to inspire and develop the package: Ian Goodfellow, Yoshua Bengio, Aaron Courville, Francis Bach
(2016, ISBN:978-0262035613) Deep Learning. Terrence J. Sejnowski (2018, ISBN:978-0262038034) The Deep Learning Revolution. Grant Sanderson
(3brown1blue) <https://www.youtube.com/playlist?list=PLZHQObOWTQDNU6R1_67000Dx_ZCJB-3pi> Neural Networks YouTube playlist. Michael A. Nielsen
<http://neuralnetworksanddeeplearning.com/> Neural Networks and Deep Learning
\enumerate{
    \item Ian Goodfellow, Yoshua Bengio, Aaron Courville, Francis Bach. Deep Learning. (2016)
    \item Terrence J. Sejnowski. The Deep Learning Revolution (The MIT Press). (2018)
    \item Neural Networks YouTube playlist by 3brown1blue: \url{https://www.youtube.com/playlist?list=PLZHQObOWTQDNU6R1_67000Dx_ZCJB-3pi}
    \item{http://neuralnetworksanddeeplearning.com/}
}
}
}
\author{
Benjamin Taylor, Department of Medicine, Lancaster University
}
\keyword{package}
