% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miscellaneous.R
\name{download_mnist}
\alias{download_mnist}
\title{download_mnist function}
\usage{
download_mnist(fn)
}
\arguments{
\item{fn}{the name of the file to save as}
}
\value{
a list, the elements of which are the sums of the elements of the arguments x and y.
}
\description{
A function to download mnist data in .RData format. File includes objects train_set, truth, test_set and test_truth
}
\examples{

# Don't run at R check because the file is large (23Mb)
# download_mnist("mnist.RData")


}
\references{
\enumerate{
    \item Ian Goodfellow, Yoshua Bengio, Aaron Courville, Francis Bach. Deep Learning. (2016)
    \item Terrence J. Sejnowski. The Deep Learning Revolution (The MIT Press). (2018)
    \item Neural Networks YouTube playlist by 3brown1blue: \url{https://www.youtube.com/playlist?list=PLZHQObOWTQDNU6R1_67000Dx_ZCJB-3pi}
    \item{http://neuralnetworksanddeeplearning.com/}
    \item{Y. LeCun, L. Bottou, Y. Bengio, and P. Haffner. "Gradient-based learning applied to document recognition." Proceedings of the IEEE, 86(11):2278-2324, November 1998}
    \item{http://yann.lecun.com/exdb/mnist/}
}
}
\seealso{
\link{network}, \link{train}, \link{backprop_evaluate}, \link{MLP_net}, \link{backpropagation_MLP}
}
