% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_functions.R
\name{multinomial}
\alias{multinomial}
\title{multinomial function}
\usage{
multinomial()
}
\value{
a list object with elements that are functions, evaluating the loss and the derivative
}
\description{
A function to evaluate the multinomial loss function and the derivative of this function to be used when training a neural network.
}
\examples{

\dontrun{
netwts <- train(dat=train_set,
                truth=truth,
                net=net,
                eps=0.001,
                tol=0.95,
                loss=multinomial(),
                batchsize=100)
}

}
\references{
\enumerate{
    \item Ian Goodfellow, Yoshua Bengio, Aaron Courville, Francis Bach. Deep Learning. (2016)
    \item Terrence J. Sejnowski. The Deep Learning Revolution (The MIT Press). (2018)
    \item Neural Networks YouTube playlist by 3brown1blue: \url{https://www.youtube.com/playlist?list=PLZHQObOWTQDNU6R1_67000Dx_ZCJB-3pi}
    \item{http://neuralnetworksanddeeplearning.com/}
}
}
\seealso{
\link{network}, \link{train}, \link{backprop_evaluate}, \link{MLP_net}, \link{backpropagation_MLP},
\link{Qloss}, \link{no_regularisation}, \link{L1_regularisation}, \link{L2_regularisation}
}
