% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_criteria.R
\name{stopping.maxit}
\alias{stopping.maxit}
\title{stopping.maxit function}
\usage{
\method{stopping}{maxit}(cost, curcost, count, tol, ...)
}
\arguments{
\item{cost}{the value of the loss function passed in}

\item{curcost}{current measure of cost, can be different to the parameter 'cost' above e.g. may consider smoothed cost over the last k iterations}

\item{count}{iteration count}

\item{tol}{tolerance, or limit}

\item{...}{additional arguments}
}
\value{
...
}
\description{
A function to halt computation when the number of iterations reaches a given threshold, tol
}
