% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_criteria.R
\name{stopping}
\alias{stopping}
\title{stopping function}
\usage{
stopping(...)
}
\arguments{
\item{...}{additional arguments}
}
\value{
method stopping
}
\description{
Generic function for implementing stopping methods
}
\seealso{
\link{stopping.default}, \link{stopping.maxit}
}
