% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_criteria.R
\name{updateStopping}
\alias{updateStopping}
\title{updateStopping function}
\usage{
updateStopping(...)
}
\arguments{
\item{...}{additional arguments}
}
\value{
method updateStopping
}
\description{
Generic function for updating stopping criteria
}
\seealso{
\link{updateStopping.classification}, \link{updateStopping.regression}
}
