% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopping_criteria.R
\name{updateStopping.regression}
\alias{updateStopping.regression}
\title{updateStopping.regression function}
\usage{
\method{updateStopping}{regression}(dat, parms, net, truth, testoutput,
  count, monitor, mx, curcost, ...)
}
\arguments{
\item{dat}{data object}

\item{parms}{model parameters}

\item{net}{an object of class network}

\item{truth}{the truth, to be compared with network outputs}

\item{testoutput}{a vector, the history of the stopping criteria}

\item{count}{iteration number}

\item{monitor}{logical, whether to produce a diagnostic plot}

\item{mx}{a number to be monitored e.g. the cost of the best performing paramerer configuration to date}

\item{curcost}{current measure of cost, can be different to the value of the loss function e.g. may consider smoothed cost (i.e. loss) over the last k iterations}

\item{...}{additional arguments}
}
\value{
curcost, testoutput and mx, used for iterating the maximisation process
}
\description{
A function to update the stopping criteria for a classification problem.
}
